+++ PluginsAndFeatures/azure-toolkit-for-intellij/azure-intellij-resource-connector-lib/src/main/java/com/microsoft/azure/toolkit/intellij/connector/function/ManagedIdentityFunctionSupported.java
@@ -1,0 +1,14 @@
+/*
+ * Copyright (c) Microsoft Corporation. All rights reserved.
+ * Licensed under the MIT License. See License.txt in the project root for license information.
+ */
+package com.microsoft.azure.toolkit.intellij.connector.function;
+
+import com.microsoft.azure.toolkit.intellij.connector.Connection;
+
+import javax.annotation.Nonnull;
+import java.util.Map;
+
+public interface ManagedIdentityFunctionSupported<T> extends FunctionSupported<T> {
+    Map<String, String> getPropertiesForIdentityFunction(@Nonnull Connection<?, ?> connection);
+}